Move ssh server setup into testing package
