+++ src/scitacean/testing/ssh/docker-compose-ssh-server.yaml
@@ -16,7 +16,7 @@ services:
       - PASSWORD_ACCESS=true
       - USER_NAME=the-scitacean
       - USER_PASSWORD=sup3r-str0ng
-    volumes:
+    volumes: # configured in Python
       - ${DATA_DIR}:/data
       - ${SEED_DIR}:/data/seed
     ports:
