+++ tests/conftest.py
@@ -5,17 +5,13 @@ import hypothesis
 import pytest
 
 from scitacean.testing.backend import add_pytest_option as add_backend_option
+from scitacean.testing.ssh import add_pytest_option as add_ssh_option
 
-from .common.ssh_server import (  # noqa: F401
-    ssh_access,
-    ssh_config_dir,
-    ssh_connect_with_username_password,
-    ssh_connection_config,
-    ssh_data_dir,
-    ssh_fileserver,
+pytest_plugins = (
+    "scitacean.testing.backend.fixtures",
+    "scitacean.testing.ssh.fixtures",
 )
 
-pytest_plugins = "scitacean.testing.backend.fixtures"
 
 # The datasets strategy requires a large amount of memory and time.
 # This is not good but hard to avoid.
@@ -31,9 +27,4 @@ hypothesis.settings.register_profile(
 
 def pytest_addoption(parser: pytest.Parser):
     add_backend_option(parser)
-    parser.addoption(
-        "--ssh-tests",
-        action="store_true",
-        default=False,
-        help="Select whether to run tests with an SSH fileserver",
-    )
+    add_ssh_option(parser)
